FE-247: General fixes
